add image variable to the user and add a pop up for account deletion

+++ client/src/views/EditPassword.vue
@@ -17,19 +17,19 @@
                             <div class="mb-3 text-start">
                                 <label class="small mb-1" for="current_password">Current Password</label>
                                 <input class="form-control"  id="current_password" type="password" v-model="current_password" placeholder="Enter current password">
-                                <div v-if="getError('current_password')" class="paint-red">{{ getError('current_password') }} </div>
+                                <div v-if="getError('current_password')" class="invalid-feedback d-block">{{ getError('current_password') }} </div>
                               </div>
                             <!-- Form Group (new password)-->
                             <div class="mb-3 text-start">
                                 <label class="small mb-1" for="new_password">New Password</label>
-                                <input class="form-control" id="new_password" type="new_password" v-model="new_password" placeholder="Enter new password">
-                                <div v-if="getError('new_password')" class="paint-red">{{ getError('new_password') }} </div>
+                                <input class="form-control" id="new_password" type="password" v-model="new_password" placeholder="Enter new password">
+                                <div v-if="getError('new_password')" class="invalid-feedback d-block">{{ getError('new_password') }} </div>
                               </div>
                             <!-- Form Group (confirm password)-->
                             <div class="mb-3 text-start">
                                 <label class="small mb-1" for="confirm_password">Confirm Password</label>
-                                <input class="form-control" id="confirm_password" type="confirm_password" v-model="confirm_password"  placeholder="Confirm new password">
-                                <div v-if="getError('confirm_password')" class="paint-red">{{ getError('confirm_password') }} </div>
+                                <input class="form-control" :class="{ 'is-invalid': getError('password_confirmation') }" id="confirm_password" type="password" v-model="confirm_password"  placeholder="Confirm new password"/>
+                                <div v-if="getError('confirm_password')"  class="invalid-feedback d-block">{{ getError('confirm_password') }} </div>
                               </div>
                             <button class="btn btn-primary" type="submit">Save</button>
                         </form>
@@ -117,7 +117,8 @@ export default {
       }
     },
     async deleteAccount() {
-      console.log('delete')
+      const text = 'Are you sure you want to delete your account?'
+      if (confirm(text) === false) return
       const user = JSON.parse(localStorage.getItem('user-info'))
       const userId = user.body._id
       const response = await Api.delete(`http://localhost:3000/v1/users/${userId}`)
@@ -163,8 +164,7 @@ export default {
     background-color: rgba(33, 40, 50, 0.03);
     border-bottom: 1px solid rgba(33, 40, 50, 0.125);
 }
-.form-control, .dataTable-input {
-    display: block;
+.form-control{
     width: 100%;
     padding: 0.875rem 1.125rem;
     font-size: 0.875rem;
